+++ sge_lib/sge/scene/sceneimporter.cpp
@@ -11,7 +11,7 @@ namespace SGE
 			return NULL;
 		
 		fseek(file, 0, SEEK_END);
-		int count = ftell(file);
+		size_t count = ftell(file);
 		rewind(file);
